refactor(app): add the 'listen' and 'getDbDriver' methods to the returned module API
Add the 'listen' and 'getDbDriver' methods to the returned module API and let users call on those
methods to bind the user-management app to their routes (using listen()) and get a DB driver of
their choice (using getDbDriver()). This is in contrast to the previous method of returning 'listen'
as the API which when called then returns the UserModule object, and also the previous manual way of
having to try to get to the db object directly by 'require()'-ing it yourself. Rename the lib/stores
directory to databases/drivers directory.

BREAKING CHANGE: Code using the previous 'listen()' will fail since the userModule instance is no
longer returned as a result of a call to 'listen()' rather it is now exported from the index.js
file, and the call to userModule.listen() returns undefined.